pyproject.toml
[tool.poetry]
name = "nanika-bot"
version = "0.1.0"
description = ""
authors = ["nanika2 <ika2kki@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "^0.29.0"
"discord.py" = { git = "https://github.com/Rapptz/discord.py.git", extras = [
    "voice",
    "speedup",
], rev = "master" }
typing-extensions = "^4.8.0"
pillow = "^10.1.0"
tabulate = "^0.9.0"
watchdog = "^3.0.0"
cryptography = "^41.0.7"
wavelink = {git = "https://github.com/PythonistaGuild/Wavelink"}
forbiddenfruit = "^0.1.4"
jishaku = {git = "https://github.com/Gorialis/jishaku"}
sphinx = "^7.2.6"
rapidfuzz = "^3.6.1"
starlight-dpy = {git = "https://github.com/InterStella0/starlight-dpy"}

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
